Create Major_Lab_1_Frequency_Of_Word_In_Sentence.jupyterlite.ipynb
First major lab to create a class that will analyze a string of text, remove punctuation and convert to lower case.

Once this is complete the class allows for the string to get a count of the frequency words are used.

Lastly you can then see what individual words freqency are in this string.